Moved search menu to the fragment
